+++ 03_ddl/044.ext_tpcds.store_sales.sql
@@ -2,7 +2,7 @@ CREATE EXTERNAL TABLE ext_tpcds.store_sales
 (
     ss_sold_date_sk integer,
     ss_sold_time_sk integer,
-    ss_item_sk bigint,
+    ss_item_sk int,
     ss_customer_sk integer,
     ss_cdemo_sk integer,
     ss_hdemo_sk integer,
